requirements.txt
black
bs4
requests
tinydb
pyyaml